Refactor of code. Code split TaskErrors to mod.rs, Serialization to persistence.rs and Tasks struct to crud.rs
